+++ guessit/rules/properties/size.py
@@ -8,7 +8,7 @@ import re
 from rebulk import Rebulk
 
 from ..common import dash
-from ..common.quantity import Quantity
+from ..common.quantity import Size
 from ..common.validators import seps_surround
 
 
@@ -20,6 +20,6 @@ def size():
     """
     rebulk = Rebulk().regex_defaults(flags=re.IGNORECASE, abbreviations=[dash])
     rebulk.defaults(name='size', validator=seps_surround)
-    rebulk.regex(r'\d+\.?[mgt]b', r'\d+\.\d+[mgt]b', formatter=Quantity.fromstring, tags=['release-group-prefix'])
+    rebulk.regex(r'\d+\.?[mgt]b', r'\d+\.\d+[mgt]b', formatter=Size.fromstring, tags=['release-group-prefix'])
 
     return rebulk